Optimises TypeBadge Component
